Adding Zoons And Stocks Connect Stocks And Zoons With Users & Products Connect Orders With Zoon Adding Close Order API

+++ LowCost.Business/Services/User/Implementation/Dashboard/DashboardDriverService.cs
@@ -4,6 +4,7 @@ using LowCost.Business.Services.User.Interfaces.Dashboard;
 using LowCost.Infrastructure.DashboardViewModels.User;
 using LowCost.Infrastructure.Helpers;
 using LowCost.Infrastructure.Pagination;
+using LowCost.Repo.UnitOfWork;
 using Microsoft.AspNetCore.Identity;
 using System;
 using System.Collections.Generic;
@@ -15,16 +16,20 @@ namespace LowCost.Business.Services.User.Implementation.Dashboard
 {
     public class DashboardDriverService : IDashboardDriverService
     {
+        private readonly IUnitOfWork _unitOfWork;
         private readonly UserManager<Domain.Models.User> _userManager;
         private readonly IMapper _mapper;
         private readonly AuthenticationHandler _authenticationHandler;
+        Domain.Models.User currentAdmin = null;
 
-        public DashboardDriverService(UserManager<Domain.Models.User> userManager,
+        public DashboardDriverService(IUnitOfWork unitOfWork, UserManager<Domain.Models.User> userManager,
             IMapper mapper, AuthenticationHandler authenticationHandler)
         {
+            this._unitOfWork = unitOfWork;
             this._userManager = userManager;
             this._mapper = mapper;
             this._authenticationHandler = authenticationHandler;
+            currentAdmin =  _unitOfWork.UsersRepository.GetCurrentDashboardAdminUser().Result;
         }
         public async Task<ActionState> CreateDriverAsync(AddDriverViewModel addDriverViewModel)
         {
@@ -40,7 +45,8 @@ namespace LowCost.Business.Services.User.Implementation.Dashboard
                 FullName = addDriverViewModel.FullName,
                 UserName = addDriverViewModel.Email,
                 Email = addDriverViewModel.Email,
-                PhoneNumber = addDriverViewModel.Phone
+                PhoneNumber = addDriverViewModel.Phone,
+                Stock_Id = currentAdmin.Stock_Id == null ? addDriverViewModel.Stock_Id : currentAdmin.Stock_Id
             };
             var result = await _userManager.CreateAsync(driver, addDriverViewModel.Password);
             if(result.Succeeded)
@@ -56,6 +62,10 @@ namespace LowCost.Business.Services.User.Implementation.Dashboard
         public async Task<IdentityResult> DeleteDriverAsync(string Id)
         {
             var driver = await _userManager.FindByIdAsync(Id);
+            if(currentAdmin.Stock_Id != null && currentAdmin.Stock_Id != driver.Stock_Id)
+            {
+                return new IdentityResult();
+            }
             if (driver != null && await _userManager.IsInRoleAsync(driver, Constants.DriverRoleName))
             {
                 return await _userManager.DeleteAsync(driver);
@@ -65,7 +75,7 @@ namespace LowCost.Business.Services.User.Implementation.Dashboard
 
         public async Task<IEnumerable<DriverViewModel>> GetAllDriversAsync()
         {
-            var drivers = await _userManager.GetUsersInRoleAsync(Constants.DriverRoleName);
+            var drivers = (await _userManager.GetUsersInRoleAsync(Constants.DriverRoleName)).Where(driver => currentAdmin.Stock_Id == null || currentAdmin.Stock_Id == driver.Stock_Id);
 
             var driversDTOs = _mapper.Map<IEnumerable<Domain.Models.User>, IEnumerable<DriverViewModel>>(drivers);
 
@@ -75,7 +85,7 @@ namespace LowCost.Business.Services.User.Implementation.Dashboard
         public async Task<PagedResult<DriverViewModel>> GetDriversAsync(PagingParameters pagingParameters)
         {
             int skip = pagingParameters.Index * pagingParameters.Size;
-            var drivers = (await _userManager.GetUsersInRoleAsync(Constants.DriverRoleName)).Skip(skip).Take(pagingParameters.Size);
+            var drivers = (await _userManager.GetUsersInRoleAsync(Constants.DriverRoleName)).Where(driver => currentAdmin.Stock_Id == null || currentAdmin.Stock_Id == driver.Stock_Id).Skip(skip).Take(pagingParameters.Size);
 
             var driversDTOs = _mapper.Map<IEnumerable<Domain.Models.User>, IEnumerable<DriverViewModel>>(drivers);
             var result = new PagedResult<DriverViewModel>()
